:children_crossing: 첫 실행시 command line tool, brew, git, jq 최대한 자동화하는 것 시도 및 설치 안되어있는 경우에만 설치하도록 수정

+++ install.sh
@@ -2,6 +2,19 @@
 
 source util.sh
 
+function install_command_line_tool() {
+    xcode-select --install
+    sleep 1
+osascript <<EOD
+  tell application "System Events"
+    tell process "Install Command Line Developer Tools"
+      keystroke return
+      click button "Agree" of window "License Agreement"
+    end tell
+  end tell
+EOD
+}
+
 function install_brew_with_expect_pw() {
     local pw=$1
     local cmd=$([ X`which brew` = X/usr/local/bin/brew ] || /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)")
@@ -26,24 +39,45 @@ function prepare_install() {
     local config_path=$1
     local pw=$2
     
-    # 설치를 위해 필요한 것들을 설치한다. (brew, git, jq)
+    # 설치를 위해 필요한 것들을 설치한다. (xcode command line developer tool, brew, git, jq)
+    # install xcode command line developer tool
+    if softwareupdate --history | grep --silent "Command Line Tools"; then
+        echo "Command line tools already installed"
+    else
+        install_command_line_tool
+    fi
+    
     # install Homebrew
-    #install_brew_with_expect_pw $pw
-    # 왜인지 잘 모르겠지만 expect 가 안먹는다 ... 별 수 없이 처음에는 비번 넣고 엔터키 쳐준다.
-    [ X`which brew` = X/usr/local/bin/brew ] || /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
+    is_brew_exist=$(is_exist_cmd brew)
+    if [[ $is_brew_exist != "exist" ]]; then
+        install_brew_with_expect_pw $pw
+        # 왜인지 잘 모르겠지만 expect 가 안먹는다 ... 별 수 없이 처음에는 비번 넣고 엔터키 쳐준다.
+        #[ X`which brew` = X/usr/local/bin/brew ] || /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install.sh)"
+    fi
+
+
 
     # install git
-    brew install git
+    is_git_exist=$(is_exist_cmd git)
+    if [[ $is_git_exist != "exist" ]]; then
+        brew install git
+    fi
 
     # install jq
-    brew install jq
+    is_jq_exist=$(is_exist_cmd jq)
+    if [[ $is_jq_exist != "exist" ]]; then
+        brew install jq
+    fi
 
     # if order contains mas -> install mas
     order=`cat $config_path | jq -r ".order[]"`
     for o in $order; do
         if [[ $o == "mas" ]] ; then
             # install mas
-            brew install mas
+            is_mas_exist=$(is_exist_cmd mas)
+            if [[ $is_mas_exist != "exist" ]]; then
+                brew install mas
+            fi
             request_login_mas
         fi
     done
